requirements.txt
flask
pywebview
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
pyscard
PyYAML